refactor: split admin-view scaffold modules
