+++ packages/wp-typia-project-tools/src/runtime/package-versions.ts
@@ -1,40 +1,40 @@
-import fs from "node:fs";
-import { createRequire } from "node:module";
-import path from "node:path";
+import fs from 'node:fs';
+import { createRequire } from 'node:module';
+import path from 'node:path';
 
-import { PROJECT_TOOLS_PACKAGE_ROOT } from "./template-registry.js";
+import { PROJECT_TOOLS_PACKAGE_ROOT } from './template-registry.js';
 
 interface PackageManifest {
-	dependencies?: Record<string, string>;
-	devDependencies?: Record<string, string>;
-	version?: string;
+  dependencies?: Record<string, string>;
+  devDependencies?: Record<string, string>;
+  version?: string;
 }
 
 export interface PackageVersions {
-	apiClientPackageVersion: string;
-	blockRuntimePackageVersion: string;
-	blockTypesPackageVersion: string;
-	projectToolsPackageVersion: string;
-	restPackageVersion: string;
-	tsxPackageVersion: string;
-	typiaPackageVersion: string;
-	typiaUnpluginPackageVersion: string;
-	typescriptPackageVersion: string;
-	wpTypiaPackageExactVersion: string;
-	wpTypiaPackageVersion: string;
+  apiClientPackageVersion: string;
+  blockRuntimePackageVersion: string;
+  blockTypesPackageVersion: string;
+  projectToolsPackageVersion: string;
+  restPackageVersion: string;
+  tsxPackageVersion: string;
+  typiaPackageVersion: string;
+  typiaUnpluginPackageVersion: string;
+  typescriptPackageVersion: string;
+  wpTypiaPackageExactVersion: string;
+  wpTypiaPackageVersion: string;
 }
 
 interface PackageManifestLocation {
-	cacheKey: string;
-	packageJsonPath: string | null;
-	source: string | null;
+  cacheKey: string;
+  packageJsonPath: string | null;
+  source: string | null;
 }
 
 const require = createRequire(import.meta.url);
-const DEFAULT_VERSION_RANGE = "^0.0.0";
-const DEFAULT_EXACT_VERSION = "0.0.0";
-const DEFAULT_TSX_PACKAGE_VERSION = "^4.20.5";
-const DEFAULT_TYPIA_UNPLUGIN_PACKAGE_VERSION = "^12.0.1";
+const DEFAULT_VERSION_RANGE = '^0.0.0';
+const DEFAULT_EXACT_VERSION = '0.0.0';
+const DEFAULT_TSX_PACKAGE_VERSION = '^4.20.5';
+const DEFAULT_TYPIA_UNPLUGIN_PACKAGE_VERSION = '^12.0.1';
 /**
  * Explicit fallback ranges for managed WordPress-facing workspace dependencies.
  *
@@ -42,115 +42,160 @@ const DEFAULT_TYPIA_UNPLUGIN_PACKAGE_VERSION = "^12.0.1";
  * fresher local or installed manifest version first, so add-command defaults do
  * not drift across runtime modules.
  */
-export const DEFAULT_WORDPRESS_ABILITIES_VERSION = "^0.10.0";
-export const DEFAULT_WORDPRESS_CORE_ABILITIES_VERSION = "^0.9.0";
-export const DEFAULT_WORDPRESS_CORE_DATA_VERSION = "^7.44.0";
-export const DEFAULT_WORDPRESS_DATA_VERSION = "^9.28.0";
-export const DEFAULT_WORDPRESS_DATAVIEWS_VERSION = "^14.1.0";
-export const DEFAULT_WP_TYPIA_DATAVIEWS_VERSION = "^0.1.0";
+export const DEFAULT_WORDPRESS_ABILITIES_VERSION = '^0.10.0';
+export const DEFAULT_WORDPRESS_CORE_ABILITIES_VERSION = '^0.9.0';
+export const DEFAULT_WORDPRESS_CORE_DATA_VERSION = '^7.44.0';
+export const DEFAULT_WORDPRESS_DATA_VERSION = '^9.28.0';
+export const DEFAULT_WORDPRESS_DATAVIEWS_VERSION = '^14.1.0';
+export const DEFAULT_WP_TYPIA_DATAVIEWS_VERSION = '^0.1.0';
 
-let cachedPackageVersions:
-	| {
-			cacheKey: string;
-			versions: PackageVersions;
-	  }
-	| null = null;
+let cachedPackageVersions: {
+  cacheKey: string;
+  versions: PackageVersions;
+} | null = null;
 
 function getErrorCode(error: unknown): string | undefined {
-	return typeof error === "object" && error !== null && "code" in error
-		? String((error as { code: unknown }).code)
-		: undefined;
+  return typeof error === 'object' && error !== null && 'code' in error
+    ? String((error as { code: unknown }).code)
+    : undefined;
 }
 
 function normalizeVersionRange(value: string | undefined): string {
-	const trimmed = value?.trim();
-	if (!trimmed) {
-		return DEFAULT_VERSION_RANGE;
-	}
-	if (trimmed.startsWith("workspace:")) {
-		return DEFAULT_VERSION_RANGE;
-	}
+  const trimmed = value?.trim();
+  if (!trimmed) {
+    return DEFAULT_VERSION_RANGE;
+  }
+  if (trimmed.startsWith('workspace:')) {
+    return DEFAULT_VERSION_RANGE;
+  }
 
-	return /^[~^<>=]/.test(trimmed) ? trimmed : `^${trimmed}`;
+  return /^[~^<>=]/.test(trimmed) ? trimmed : `^${trimmed}`;
 }
 
 function normalizeExactVersion(value: string | undefined): string {
-	const trimmed = value?.trim();
-	if (!trimmed) {
-		return DEFAULT_EXACT_VERSION;
-	}
-	if (trimmed.startsWith("workspace:")) {
-		return DEFAULT_EXACT_VERSION;
-	}
-	return trimmed.replace(/^[~^<>=]+/, "");
+  const trimmed = value?.trim();
+  if (!trimmed) {
+    return DEFAULT_EXACT_VERSION;
+  }
+  if (trimmed.startsWith('workspace:')) {
+    return DEFAULT_EXACT_VERSION;
+  }
+  return trimmed.replace(/^[~^<>=]+/, '');
 }
 
 function normalizeVersionRangeWithFallback(
-	value: string | undefined,
-	fallback: string,
+  value: string | undefined,
+  fallback: string,
 ): string {
-	const normalized = normalizeVersionRange(value);
-	return normalized === DEFAULT_VERSION_RANGE ? fallback : normalized;
+  const normalized = normalizeVersionRange(value);
+  return normalized === DEFAULT_VERSION_RANGE ? fallback : normalized;
 }
 
 function createContentFingerprint(source: string): string {
-	let hash = 2166136261;
-	for (let index = 0; index < source.length; index += 1) {
-		hash ^= source.charCodeAt(index);
-		hash = Math.imul(hash, 16777619);
-	}
+  let hash = 2166136261;
+  for (let index = 0; index < source.length; index += 1) {
+    hash ^= source.charCodeAt(index);
+    hash = Math.imul(hash, 16777619);
+  }
 
-	return (hash >>> 0).toString(16);
+  return (hash >>> 0).toString(16);
 }
 
-function resolvePackageManifestLocation(packageJsonPath: string): PackageManifestLocation {
-	try {
-		const stats = fs.statSync(packageJsonPath);
-		const source = fs.readFileSync(packageJsonPath, "utf8");
-		return {
-			cacheKey: `file:${packageJsonPath}:${stats.ino}:${stats.mtimeMs}:${stats.ctimeMs}:${stats.size}:${createContentFingerprint(source)}`,
-			packageJsonPath,
-			source,
-		};
-	} catch (error) {
-		if (getErrorCode(error) === "ENOENT") {
-			return {
-				cacheKey: `missing-file:${packageJsonPath}`,
-				packageJsonPath: null,
-				source: null,
-			};
-		}
-		throw error;
-	}
+function resolvePackageManifestLocation(
+  packageJsonPath: string,
+): PackageManifestLocation {
+  try {
+    const stats = fs.statSync(packageJsonPath);
+    const source = fs.readFileSync(packageJsonPath, 'utf8');
+    return {
+      cacheKey: `file:${packageJsonPath}:${stats.ino}:${stats.mtimeMs}:${stats.ctimeMs}:${stats.size}:${createContentFingerprint(source)}`,
+      packageJsonPath,
+      source,
+    };
+  } catch (error) {
+    if (getErrorCode(error) === 'ENOENT') {
+      return {
+        cacheKey: `missing-file:${packageJsonPath}`,
+        packageJsonPath: null,
+        source: null,
+      };
+    }
+    throw error;
+  }
 }
 
-function readPackageManifest(location: PackageManifestLocation): PackageManifest | null {
-	if (!location.packageJsonPath || location.source === null) {
-		return null;
-	}
+function readPackageManifest(
+  location: PackageManifestLocation,
+): PackageManifest | null {
+  if (!location.packageJsonPath || location.source === null) {
+    return null;
+  }
 
-	return JSON.parse(location.source) as PackageManifest;
+  return JSON.parse(location.source) as PackageManifest;
 }
 
-function resolveInstalledPackageManifestLocation(packageName: string): PackageManifestLocation {
-	try {
-		return resolvePackageManifestLocation(require.resolve(`${packageName}/package.json`));
-	} catch (error) {
-		if (getErrorCode(error) === "MODULE_NOT_FOUND") {
-			return {
-				cacheKey: `missing-module:${packageName}`,
-				packageJsonPath: null,
-				source: null,
-			};
-		}
-		throw error;
-	}
+function resolveInstalledPackageManifestLocation(
+  packageName: string,
+): PackageManifestLocation {
+  try {
+    return resolvePackageManifestLocation(
+      require.resolve(`${packageName}/package.json`),
+    );
+  } catch (error) {
+    if (getErrorCode(error) === 'MODULE_NOT_FOUND') {
+      return {
+        cacheKey: `missing-module:${packageName}`,
+        packageJsonPath: null,
+        source: null,
+      };
+    }
+    throw error;
+  }
 }
 
 function composePackageVersionsCacheKey(
-	locations: readonly PackageManifestLocation[],
+  locations: readonly PackageManifestLocation[],
 ): string {
-	return locations.map((location) => location.cacheKey).join("|");
+  return locations.map((location) => location.cacheKey).join('|');
+}
+
+/**
+ * Resolve a managed package version range from linked workspace packages first,
+ * then installed package manifests, while preserving the shared normalization
+ * and manifest fingerprinting rules used by `getPackageVersions()`.
+ *
+ * @param packageName npm package whose manifest version should be consulted.
+ * @param fallback Canonical range to use when no usable manifest version exists.
+ * @param workspacePackageDirName Optional sibling monorepo package directory name.
+ * @returns A normalized semver range suitable for generated dependency entries.
+ */
+export function resolveManagedPackageVersionRange(options: {
+  fallback: string;
+  packageName: string;
+  workspacePackageDirName?: string;
+}): string {
+  const workspaceManifestLocation = options.workspacePackageDirName
+    ? resolvePackageManifestLocation(
+        path.join(
+          PROJECT_TOOLS_PACKAGE_ROOT,
+          '..',
+          options.workspacePackageDirName,
+          'package.json',
+        ),
+      )
+    : null;
+  const installedManifestLocation = resolveInstalledPackageManifestLocation(
+    options.packageName,
+  );
+  const workspaceManifest = workspaceManifestLocation
+    ? readPackageManifest(workspaceManifestLocation)
+    : null;
+  const installedManifest = readPackageManifest(installedManifestLocation);
+
+  return normalizeVersionRangeWithFallback(
+    workspaceManifest?.version ?? installedManifest?.version,
+    options.fallback,
+  );
 }
 
 /**
@@ -161,7 +206,7 @@ function composePackageVersionsCacheKey(
  * synchronously from disk.
  */
 export function clearPackageVersionsCache(): void {
-	cachedPackageVersions = null;
+  cachedPackageVersions = null;
 }
 
 /**
@@ -169,7 +214,7 @@ export function clearPackageVersionsCache(): void {
  * internal invalidation name before the public cache policy was documented.
  */
 export function invalidatePackageVersionsCache(): void {
-	clearPackageVersionsCache();
+  clearPackageVersionsCache();
 }
 
 /**
@@ -180,127 +225,132 @@ export function invalidatePackageVersionsCache(): void {
  * disk, the cache key changes and the returned version object is refreshed.
  */
 export function getPackageVersions(): PackageVersions {
-	const createManifestLocation = resolvePackageManifestLocation(
-		path.join(PROJECT_TOOLS_PACKAGE_ROOT, "package.json"),
-	);
-	const monorepoManifestLocation = resolvePackageManifestLocation(
-		path.join(PROJECT_TOOLS_PACKAGE_ROOT, "..", "..", "package.json"),
-	);
-	const blockRuntimeManifestLocation = resolvePackageManifestLocation(
-		path.join(PROJECT_TOOLS_PACKAGE_ROOT, "..", "wp-typia-block-runtime", "package.json"),
-	);
-	const wpTypiaManifestLocation = resolvePackageManifestLocation(
-		path.join(PROJECT_TOOLS_PACKAGE_ROOT, "..", "wp-typia", "package.json"),
-	);
-	const installedProjectToolsManifestLocation =
-		resolveInstalledPackageManifestLocation("@wp-typia/project-tools");
-	const installedApiClientManifestLocation =
-		resolveInstalledPackageManifestLocation("@wp-typia/api-client");
-	const installedBlockRuntimeManifestLocation =
-		resolveInstalledPackageManifestLocation("@wp-typia/block-runtime");
-	const installedBlockTypesManifestLocation =
-		resolveInstalledPackageManifestLocation("@wp-typia/block-types");
-	const installedRestManifestLocation =
-		resolveInstalledPackageManifestLocation("@wp-typia/rest");
-	const installedTsxManifestLocation =
-		resolveInstalledPackageManifestLocation("tsx");
-	const installedTypiaManifestLocation =
-		resolveInstalledPackageManifestLocation("typia");
-	const installedTypiaUnpluginManifestLocation =
-		resolveInstalledPackageManifestLocation("@typia/unplugin");
-	const installedTypescriptManifestLocation =
-		resolveInstalledPackageManifestLocation("typescript");
-	const installedWpTypiaManifestLocation =
-		resolveInstalledPackageManifestLocation("wp-typia");
-	const cacheKey = composePackageVersionsCacheKey([
-		createManifestLocation,
-		monorepoManifestLocation,
-		blockRuntimeManifestLocation,
-		wpTypiaManifestLocation,
-		installedProjectToolsManifestLocation,
-		installedApiClientManifestLocation,
-		installedBlockRuntimeManifestLocation,
-		installedBlockTypesManifestLocation,
-		installedRestManifestLocation,
-		installedTsxManifestLocation,
-		installedTypiaManifestLocation,
-		installedTypiaUnpluginManifestLocation,
-		installedTypescriptManifestLocation,
-		installedWpTypiaManifestLocation,
-	]);
+  const createManifestLocation = resolvePackageManifestLocation(
+    path.join(PROJECT_TOOLS_PACKAGE_ROOT, 'package.json'),
+  );
+  const monorepoManifestLocation = resolvePackageManifestLocation(
+    path.join(PROJECT_TOOLS_PACKAGE_ROOT, '..', '..', 'package.json'),
+  );
+  const blockRuntimeManifestLocation = resolvePackageManifestLocation(
+    path.join(
+      PROJECT_TOOLS_PACKAGE_ROOT,
+      '..',
+      'wp-typia-block-runtime',
+      'package.json',
+    ),
+  );
+  const wpTypiaManifestLocation = resolvePackageManifestLocation(
+    path.join(PROJECT_TOOLS_PACKAGE_ROOT, '..', 'wp-typia', 'package.json'),
+  );
+  const installedProjectToolsManifestLocation =
+    resolveInstalledPackageManifestLocation('@wp-typia/project-tools');
+  const installedApiClientManifestLocation =
+    resolveInstalledPackageManifestLocation('@wp-typia/api-client');
+  const installedBlockRuntimeManifestLocation =
+    resolveInstalledPackageManifestLocation('@wp-typia/block-runtime');
+  const installedBlockTypesManifestLocation =
+    resolveInstalledPackageManifestLocation('@wp-typia/block-types');
+  const installedRestManifestLocation =
+    resolveInstalledPackageManifestLocation('@wp-typia/rest');
+  const installedTsxManifestLocation =
+    resolveInstalledPackageManifestLocation('tsx');
+  const installedTypiaManifestLocation =
+    resolveInstalledPackageManifestLocation('typia');
+  const installedTypiaUnpluginManifestLocation =
+    resolveInstalledPackageManifestLocation('@typia/unplugin');
+  const installedTypescriptManifestLocation =
+    resolveInstalledPackageManifestLocation('typescript');
+  const installedWpTypiaManifestLocation =
+    resolveInstalledPackageManifestLocation('wp-typia');
+  const cacheKey = composePackageVersionsCacheKey([
+    createManifestLocation,
+    monorepoManifestLocation,
+    blockRuntimeManifestLocation,
+    wpTypiaManifestLocation,
+    installedProjectToolsManifestLocation,
+    installedApiClientManifestLocation,
+    installedBlockRuntimeManifestLocation,
+    installedBlockTypesManifestLocation,
+    installedRestManifestLocation,
+    installedTsxManifestLocation,
+    installedTypiaManifestLocation,
+    installedTypiaUnpluginManifestLocation,
+    installedTypescriptManifestLocation,
+    installedWpTypiaManifestLocation,
+  ]);
 
-	if (cachedPackageVersions?.cacheKey === cacheKey) {
-		return cachedPackageVersions.versions;
-	}
+  if (cachedPackageVersions?.cacheKey === cacheKey) {
+    return cachedPackageVersions.versions;
+  }
 
-	const createManifest =
-		readPackageManifest(createManifestLocation) ??
-		readPackageManifest(installedProjectToolsManifestLocation) ??
-		{};
-	const monorepoManifest = readPackageManifest(monorepoManifestLocation) ?? {};
-	const blockRuntimeManifest =
-		readPackageManifest(blockRuntimeManifestLocation) ??
-		readPackageManifest(installedBlockRuntimeManifestLocation) ??
-		{};
-	const wpTypiaManifest =
-		readPackageManifest(wpTypiaManifestLocation) ??
-		readPackageManifest(installedWpTypiaManifestLocation) ??
-		{};
-	const blockRuntimeDependencyVersion = normalizeVersionRange(
-		createManifest.dependencies?.["@wp-typia/block-runtime"],
-	);
-	const versions = {
-		apiClientPackageVersion: normalizeVersionRange(
-			createManifest.dependencies?.["@wp-typia/api-client"] ??
-				readPackageManifest(installedApiClientManifestLocation)?.version,
-		),
-		blockRuntimePackageVersion:
-			blockRuntimeDependencyVersion !== DEFAULT_VERSION_RANGE
-				? blockRuntimeDependencyVersion
-				: normalizeVersionRange(blockRuntimeManifest.version),
-		blockTypesPackageVersion: normalizeVersionRange(
-			createManifest.dependencies?.["@wp-typia/block-types"] ??
-				readPackageManifest(installedBlockTypesManifestLocation)?.version,
-		),
-		projectToolsPackageVersion: normalizeVersionRange(createManifest.version),
-		restPackageVersion: normalizeVersionRange(
-			createManifest.dependencies?.["@wp-typia/rest"] ??
-				readPackageManifest(installedRestManifestLocation)?.version,
-		),
-		tsxPackageVersion: normalizeVersionRangeWithFallback(
-			monorepoManifest.dependencies?.tsx ??
-				monorepoManifest.devDependencies?.tsx ??
-				readPackageManifest(installedTsxManifestLocation)?.version,
-			DEFAULT_TSX_PACKAGE_VERSION,
-		),
-		typiaPackageVersion: normalizeVersionRangeWithFallback(
-			monorepoManifest.dependencies?.typia ??
-				monorepoManifest.devDependencies?.typia ??
-				createManifest.dependencies?.typia ??
-				readPackageManifest(installedTypiaManifestLocation)?.version,
-			DEFAULT_VERSION_RANGE,
-		),
-		typiaUnpluginPackageVersion: normalizeVersionRangeWithFallback(
-			monorepoManifest.dependencies?.["@typia/unplugin"] ??
-				monorepoManifest.devDependencies?.["@typia/unplugin"] ??
-				readPackageManifest(installedTypiaUnpluginManifestLocation)?.version,
-			DEFAULT_TYPIA_UNPLUGIN_PACKAGE_VERSION,
-		),
-		typescriptPackageVersion: normalizeVersionRangeWithFallback(
-			monorepoManifest.dependencies?.typescript ??
-				monorepoManifest.devDependencies?.typescript ??
-				createManifest.dependencies?.typescript ??
-				readPackageManifest(installedTypescriptManifestLocation)?.version,
-			DEFAULT_VERSION_RANGE,
-		),
-		wpTypiaPackageExactVersion: normalizeExactVersion(wpTypiaManifest.version),
-		wpTypiaPackageVersion: normalizeVersionRange(wpTypiaManifest.version),
-	};
+  const createManifest =
+    readPackageManifest(createManifestLocation) ??
+    readPackageManifest(installedProjectToolsManifestLocation) ??
+    {};
+  const monorepoManifest = readPackageManifest(monorepoManifestLocation) ?? {};
+  const blockRuntimeManifest =
+    readPackageManifest(blockRuntimeManifestLocation) ??
+    readPackageManifest(installedBlockRuntimeManifestLocation) ??
+    {};
+  const wpTypiaManifest =
+    readPackageManifest(wpTypiaManifestLocation) ??
+    readPackageManifest(installedWpTypiaManifestLocation) ??
+    {};
+  const blockRuntimeDependencyVersion = normalizeVersionRange(
+    createManifest.dependencies?.['@wp-typia/block-runtime'],
+  );
+  const versions = {
+    apiClientPackageVersion: normalizeVersionRange(
+      createManifest.dependencies?.['@wp-typia/api-client'] ??
+        readPackageManifest(installedApiClientManifestLocation)?.version,
+    ),
+    blockRuntimePackageVersion:
+      blockRuntimeDependencyVersion !== DEFAULT_VERSION_RANGE
+        ? blockRuntimeDependencyVersion
+        : normalizeVersionRange(blockRuntimeManifest.version),
+    blockTypesPackageVersion: normalizeVersionRange(
+      createManifest.dependencies?.['@wp-typia/block-types'] ??
+        readPackageManifest(installedBlockTypesManifestLocation)?.version,
+    ),
+    projectToolsPackageVersion: normalizeVersionRange(createManifest.version),
+    restPackageVersion: normalizeVersionRange(
+      createManifest.dependencies?.['@wp-typia/rest'] ??
+        readPackageManifest(installedRestManifestLocation)?.version,
+    ),
+    tsxPackageVersion: normalizeVersionRangeWithFallback(
+      monorepoManifest.dependencies?.tsx ??
+        monorepoManifest.devDependencies?.tsx ??
+        readPackageManifest(installedTsxManifestLocation)?.version,
+      DEFAULT_TSX_PACKAGE_VERSION,
+    ),
+    typiaPackageVersion: normalizeVersionRangeWithFallback(
+      monorepoManifest.dependencies?.typia ??
+        monorepoManifest.devDependencies?.typia ??
+        createManifest.dependencies?.typia ??
+        readPackageManifest(installedTypiaManifestLocation)?.version,
+      DEFAULT_VERSION_RANGE,
+    ),
+    typiaUnpluginPackageVersion: normalizeVersionRangeWithFallback(
+      monorepoManifest.dependencies?.['@typia/unplugin'] ??
+        monorepoManifest.devDependencies?.['@typia/unplugin'] ??
+        readPackageManifest(installedTypiaUnpluginManifestLocation)?.version,
+      DEFAULT_TYPIA_UNPLUGIN_PACKAGE_VERSION,
+    ),
+    typescriptPackageVersion: normalizeVersionRangeWithFallback(
+      monorepoManifest.dependencies?.typescript ??
+        monorepoManifest.devDependencies?.typescript ??
+        createManifest.dependencies?.typescript ??
+        readPackageManifest(installedTypescriptManifestLocation)?.version,
+      DEFAULT_VERSION_RANGE,
+    ),
+    wpTypiaPackageExactVersion: normalizeExactVersion(wpTypiaManifest.version),
+    wpTypiaPackageVersion: normalizeVersionRange(wpTypiaManifest.version),
+  };
 
-	cachedPackageVersions = {
-		cacheKey,
-		versions,
-	};
+  cachedPackageVersions = {
+    cacheKey,
+    versions,
+  };
 
-	return versions;
+  return versions;
 }
