+++ packages/wp-typia-project-tools/src/runtime/string-case.ts
@@ -1,5 +1,5 @@
 function capitalizeSegment(segment: string): string {
-	return segment.charAt(0).toUpperCase() + segment.slice(1);
+  return segment.charAt(0).toUpperCase() + segment.slice(1);
 }
 
 /**
@@ -9,13 +9,13 @@ function capitalizeSegment(segment: string): string {
  * @returns A lowercase kebab-case string with collapsed separators.
  */
 export function toKebabCase(input: string): string {
-	return input
-		.trim()
-		.replace(/([a-z0-9])([A-Z])/g, "$1-$2")
-		.replace(/[^A-Za-z0-9]+/g, "-")
-		.replace(/^-+|-+$/g, "")
-		.replace(/-{2,}/g, "-")
-		.toLowerCase();
+  return input
+    .trim()
+    .replace(/([a-z0-9])([A-Z])/g, '$1-$2')
+    .replace(/[^A-Za-z0-9]+/g, '-')
+    .replace(/^-+|-+$/g, '')
+    .replace(/-{2,}/g, '-')
+    .toLowerCase();
 }
 
 /**
@@ -25,7 +25,7 @@ export function toKebabCase(input: string): string {
  * @returns A lowercase snake_case string derived from the kebab-case form.
  */
 export function toSnakeCase(input: string): string {
-	return toKebabCase(input).replace(/-/g, "_");
+  return toKebabCase(input).replace(/-/g, '_');
 }
 
 /**
@@ -35,11 +35,22 @@ export function toSnakeCase(input: string): string {
  * @returns A PascalCase string derived from the normalized kebab-case form.
  */
 export function toPascalCase(input: string): string {
-	return toKebabCase(input)
-		.split("-")
-		.filter(Boolean)
-		.map(capitalizeSegment)
-		.join("");
+  return toKebabCase(input)
+    .split('-')
+    .filter(Boolean)
+    .map(capitalizeSegment)
+    .join('');
+}
+
+/**
+ * Normalize arbitrary text into a camelCase identifier.
+ *
+ * @param input Raw text that may contain spaces, punctuation, or camelCase.
+ * @returns A camelCase string derived from the normalized PascalCase form.
+ */
+export function toCamelCase(input: string): string {
+  const pascalCase = toPascalCase(input);
+  return `${pascalCase.charAt(0).toLowerCase()}${pascalCase.slice(1)}`;
 }
 
 /**
@@ -50,13 +61,13 @@ export function toPascalCase(input: string): string {
  * @returns A PascalCase string that preserves acronyms inside segments.
  */
 export function toSegmentPascalCase(input: string): string {
-	return input
-		.replace(/[^A-Za-z0-9]+/g, " ")
-		.trim()
-		.split(/\s+/)
-		.filter(Boolean)
-		.map(capitalizeSegment)
-		.join("");
+  return input
+    .replace(/[^A-Za-z0-9]+/g, ' ')
+    .trim()
+    .split(/\s+/)
+    .filter(Boolean)
+    .map(capitalizeSegment)
+    .join('');
 }
 
 /**
@@ -66,9 +77,9 @@ export function toSegmentPascalCase(input: string): string {
  * @returns A title-cased string derived from the normalized kebab-case form.
  */
 export function toTitleCase(input: string): string {
-	return toKebabCase(input)
-		.split("-")
-		.filter(Boolean)
-		.map(capitalizeSegment)
-		.join(" ");
+  return toKebabCase(input)
+    .split('-')
+    .filter(Boolean)
+    .map(capitalizeSegment)
+    .join(' ');
 }
