+++ packages/wp-typia-project-tools/tests/package-versions.test.ts
@@ -1,58 +1,58 @@
-import { afterAll, describe, expect, test } from "bun:test";
-import { spawnSync } from "node:child_process";
-import * as fs from "node:fs";
-import * as path from "node:path";
-import { pathToFileURL } from "node:url";
+import { afterAll, describe, expect, test } from 'bun:test';
+import { spawnSync } from 'node:child_process';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { pathToFileURL } from 'node:url';
 
 import {
-	cleanupScaffoldTempRoot,
-	createScaffoldTempRoot,
-} from "./helpers/scaffold-test-harness.js";
+  cleanupScaffoldTempRoot,
+  createScaffoldTempRoot,
+} from './helpers/scaffold-test-harness.js';
 
 function writeProjectToolsManifest(
-	projectToolsRoot: string,
-	options: {
-		apiClientVersion: string;
-		projectToolsVersion: string;
-	},
+  projectToolsRoot: string,
+  options: {
+    apiClientVersion: string;
+    projectToolsVersion: string;
+  },
 ): void {
-	fs.writeFileSync(
-		path.join(projectToolsRoot, "package.json"),
-		`${JSON.stringify(
-			{
-				dependencies: {
-					"@wp-typia/api-client": options.apiClientVersion,
-				},
-				name: "@wp-typia/project-tools",
-				version: options.projectToolsVersion,
-			},
-			null,
-			2,
-		)}\n`,
-		"utf8",
-	);
+  fs.writeFileSync(
+    path.join(projectToolsRoot, 'package.json'),
+    `${JSON.stringify(
+      {
+        dependencies: {
+          '@wp-typia/api-client': options.apiClientVersion,
+        },
+        name: '@wp-typia/project-tools',
+        version: options.projectToolsVersion,
+      },
+      null,
+      2,
+    )}\n`,
+    'utf8',
+  );
 }
 
-describe("package version cache invalidation", () => {
-	const tempRoot = createScaffoldTempRoot("wp-typia-package-versions-");
+describe('package version cache invalidation', () => {
+  const tempRoot = createScaffoldTempRoot('wp-typia-package-versions-');
 
-	afterAll(() => {
-		cleanupScaffoldTempRoot(tempRoot);
-	});
+  afterAll(() => {
+    cleanupScaffoldTempRoot(tempRoot);
+  });
 
-	test("refreshes linked package metadata and exposes explicit cache clearing", () => {
-		const linkedProjectToolsRoot = fs.mkdtempSync(
-			path.join(tempRoot, "linked-project-tools-"),
-		);
-		writeProjectToolsManifest(linkedProjectToolsRoot, {
-			apiClientVersion: "^0.4.0",
-			projectToolsVersion: "1.2.3",
-		});
+  test('refreshes linked package metadata and exposes explicit cache clearing', () => {
+    const linkedProjectToolsRoot = fs.mkdtempSync(
+      path.join(tempRoot, 'linked-project-tools-'),
+    );
+    writeProjectToolsManifest(linkedProjectToolsRoot, {
+      apiClientVersion: '^0.4.0',
+      projectToolsVersion: '1.2.3',
+    });
 
-		const packageVersionsModuleUrl = pathToFileURL(
-			path.join(import.meta.dir, "..", "src", "runtime", "package-versions.ts"),
-		).href;
-		const script = `
+    const packageVersionsModuleUrl = pathToFileURL(
+      path.join(import.meta.dir, '..', 'src', 'runtime', 'package-versions.ts'),
+    ).href;
+    const script = `
 			import assert from "node:assert/strict";
 			import fs from "node:fs";
 			import path from "node:path";
@@ -101,76 +101,119 @@ describe("package version cache invalidation", () => {
 			assert.equal(refreshedByAlias.projectToolsPackageVersion, "^1.2.4");
 		`;
 
-		// The inline script imports TypeScript source, so keep the child process on Bun
-		// even if this test is ever launched through a Node-based wrapper.
-		const bunBinary =
-			process.env.BUN_BINARY ?? ("Bun" in globalThis ? process.execPath : "bun");
-		const result = spawnSync(bunBinary, ["--eval", script], {
-			cwd: linkedProjectToolsRoot,
-			encoding: "utf8",
-			env: {
-				...process.env,
-				WP_TYPIA_PROJECT_TOOLS_PACKAGE_ROOT: linkedProjectToolsRoot,
-			},
-		});
+    // The inline script imports TypeScript source, so keep the child process on Bun
+    // even if this test is ever launched through a Node-based wrapper.
+    const bunBinary =
+      process.env.BUN_BINARY ??
+      ('Bun' in globalThis ? process.execPath : 'bun');
+    const result = spawnSync(bunBinary, ['--eval', script], {
+      cwd: linkedProjectToolsRoot,
+      encoding: 'utf8',
+      env: {
+        ...process.env,
+        WP_TYPIA_PROJECT_TOOLS_PACKAGE_ROOT: linkedProjectToolsRoot,
+      },
+    });
 
-		if (result.status !== 0) {
-			throw new Error(
-				`package-versions cache script failed (status=${result.status}, error=${
-					result.error?.message ?? "none"
-				}):\n${result.stderr}`,
-			);
-		}
-		expect(result.status).toBe(0);
-		expect(result.stdout).toBe("");
-		expect(result.stderr).toBe("");
-	});
+    if (result.status !== 0) {
+      throw new Error(
+        `package-versions cache script failed (status=${result.status}, error=${
+          result.error?.message ?? 'none'
+        }):\n${result.stderr}`,
+      );
+    }
+    expect(result.status).toBe(0);
+    expect(result.stdout).toBe('');
+    expect(result.stderr).toBe('');
+  });
 
-	test("centralizes managed workspace dependency fallback ranges", async () => {
-		const packageVersionsModuleUrl = pathToFileURL(
-			path.join(import.meta.dir, "..", "src", "runtime", "package-versions.ts"),
-		).href;
-		const {
-			DEFAULT_WORDPRESS_ABILITIES_VERSION,
-			DEFAULT_WORDPRESS_CORE_ABILITIES_VERSION,
-			DEFAULT_WORDPRESS_CORE_DATA_VERSION,
-			DEFAULT_WORDPRESS_DATA_VERSION,
-			DEFAULT_WORDPRESS_DATAVIEWS_VERSION,
-			DEFAULT_WP_TYPIA_DATAVIEWS_VERSION,
-		} = await import(packageVersionsModuleUrl);
+  test('centralizes managed workspace dependency fallback ranges', async () => {
+    const packageVersionsModuleUrl = pathToFileURL(
+      path.join(import.meta.dir, '..', 'src', 'runtime', 'package-versions.ts'),
+    ).href;
+    const {
+      DEFAULT_WORDPRESS_ABILITIES_VERSION,
+      DEFAULT_WORDPRESS_CORE_ABILITIES_VERSION,
+      DEFAULT_WORDPRESS_CORE_DATA_VERSION,
+      DEFAULT_WORDPRESS_DATA_VERSION,
+      DEFAULT_WORDPRESS_DATAVIEWS_VERSION,
+      DEFAULT_WP_TYPIA_DATAVIEWS_VERSION,
+    } = await import(packageVersionsModuleUrl);
 
-		const abilityRuntimeSource = fs.readFileSync(
-			path.join(import.meta.dir, "..", "src", "runtime", "cli-add-workspace-ability.ts"),
-			"utf8",
-		);
-		const adminViewRuntimeSource = fs.readFileSync(
-			path.join(import.meta.dir, "..", "src", "runtime", "cli-add-workspace-admin-view.ts"),
-			"utf8",
-		);
+    const abilityRuntimeSource = fs.readFileSync(
+      path.join(
+        import.meta.dir,
+        '..',
+        'src',
+        'runtime',
+        'cli-add-workspace-ability.ts',
+      ),
+      'utf8',
+    );
+    const adminViewRuntimeSource = fs.readFileSync(
+      path.join(
+        import.meta.dir,
+        '..',
+        'src',
+        'runtime',
+        'cli-add-workspace-admin-view.ts',
+      ),
+      'utf8',
+    );
+    const adminViewScaffoldSource = fs.readFileSync(
+      path.join(
+        import.meta.dir,
+        '..',
+        'src',
+        'runtime',
+        'cli-add-workspace-admin-view-scaffold.ts',
+      ),
+      'utf8',
+    );
 
-		expect(DEFAULT_WORDPRESS_ABILITIES_VERSION).toBe("^0.10.0");
-		expect(DEFAULT_WORDPRESS_CORE_ABILITIES_VERSION).toBe("^0.9.0");
-		expect(DEFAULT_WORDPRESS_CORE_DATA_VERSION).toBe("^7.44.0");
-		expect(DEFAULT_WORDPRESS_DATA_VERSION).toBe("^9.28.0");
-		expect(DEFAULT_WORDPRESS_DATAVIEWS_VERSION).toBe("^14.1.0");
-		expect(DEFAULT_WP_TYPIA_DATAVIEWS_VERSION).toBe("^0.1.0");
-		expect(abilityRuntimeSource).not.toContain("const WP_ABILITIES_PACKAGE_VERSION");
-		expect(abilityRuntimeSource).not.toContain(
-			"const WP_CORE_ABILITIES_PACKAGE_VERSION",
-		);
-		expect(adminViewRuntimeSource).not.toContain(
-			"const DEFAULT_WP_TYPIA_DATAVIEWS_VERSION",
-		);
-		expect(adminViewRuntimeSource).not.toContain(
-			"const DEFAULT_WORDPRESS_DATAVIEWS_VERSION",
-		);
-		expect(adminViewRuntimeSource).not.toContain(
-			"const DEFAULT_WORDPRESS_CORE_DATA_VERSION",
-		);
-		expect(adminViewRuntimeSource).not.toContain(
-			"const DEFAULT_WORDPRESS_DATA_VERSION",
-		);
-		expect(abilityRuntimeSource).toContain('from "./package-versions.js"');
-		expect(adminViewRuntimeSource).toContain('from "./package-versions.js"');
-	});
+    expect(DEFAULT_WORDPRESS_ABILITIES_VERSION).toBe('^0.10.0');
+    expect(DEFAULT_WORDPRESS_CORE_ABILITIES_VERSION).toBe('^0.9.0');
+    expect(DEFAULT_WORDPRESS_CORE_DATA_VERSION).toBe('^7.44.0');
+    expect(DEFAULT_WORDPRESS_DATA_VERSION).toBe('^9.28.0');
+    expect(DEFAULT_WORDPRESS_DATAVIEWS_VERSION).toBe('^14.1.0');
+    expect(DEFAULT_WP_TYPIA_DATAVIEWS_VERSION).toBe('^0.1.0');
+    expect(abilityRuntimeSource).not.toContain(
+      'const WP_ABILITIES_PACKAGE_VERSION',
+    );
+    expect(abilityRuntimeSource).not.toContain(
+      'const WP_CORE_ABILITIES_PACKAGE_VERSION',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'const DEFAULT_WP_TYPIA_DATAVIEWS_VERSION',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'const DEFAULT_WORDPRESS_DATAVIEWS_VERSION',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'const DEFAULT_WORDPRESS_CORE_DATA_VERSION',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'const DEFAULT_WORDPRESS_DATA_VERSION',
+    );
+    expect(abilityRuntimeSource).toContain('from "./package-versions.js"');
+    expect(adminViewScaffoldSource).toContain("from './package-versions.js'");
+    expect(adminViewScaffoldSource).toContain(
+      'resolveManagedPackageVersionRange',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'function normalizeVersionRange(',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'function readPackageManifest(',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'function readPackageManifestVersion(',
+    );
+    expect(adminViewScaffoldSource).not.toContain(
+      'function resolvePackageVersionRange(',
+    );
+    expect(adminViewRuntimeSource).not.toContain(
+      'DEFAULT_WORDPRESS_DATAVIEWS_VERSION',
+    );
+  });
 });
